use proper integration test format
